Add clang-tidy check for overflow in constructing energy quantity from int (#71425)
* Add clang-tidy check for overflow in constructing energy quantity from int

* Avoid int overflow in calling unit::from_joule and unit::from_kilojoule

* Suppress Clang Static Analyzer temporarily

+++ src/crafting.cpp
@@ -2367,7 +2367,7 @@ void Character::consume_tools( map &m, const comp_selection<tool_comp> &tool, in
         m.use_charges( reachable_pts, tool.comp.type, quantity, return_true<item>, bcp );
         // Map::use_charges() does not handle UPS charges.
         if( quantity > 0 ) {
-            m.consume_ups( reachable_pts, units::from_kilojoule( quantity ) );
+            m.consume_ups( reachable_pts, units::from_kilojoule( static_cast<std::int64_t>( quantity ) ) );
         }
     }
